GUI Update
The GUI has been simplified. Now you can find all the cameras in the sidebar and click all of them (for now nothing interesting happens, there is only a label in the right side of the app which changes). In the main (right) part of the screen there is the currently clicked camera (the first one by default when you run the project). Files 'requirements.txt' and 'README.md' also updated.