ui and analysis update
The processed stream is now divided in a 4x4 grid (16 boxes); each box is processed alone and all the values are added into the total count of the people on the scene. The model processes a box every 15 msec, so an entire frame is processed between 2.5 and 3 seconds. You can see in the UI also the counter updated for each box.